+++ srcs/cpp/include/kungfu/python/init.h
@@ -1,9 +1,5 @@
 #pragma once
-#include <map>
 #include <memory>
-#include <mutex>
-#include <string>
-#include <vector>
 
 #include <kungfu.h>
 #include <kungfu/nccl/common.hpp>
